workflow-3/ollama/requirements.txt
paho-mqtt
ollama